# pre_award/fund_store/openapi/api.yml
openapi: "3.0.0"
info:
  title: Funding Service Design - Fund store.
  description: Fund store API for DLUHC Funding Service Design
  version: "0.2.0"
paths:
...
  /funds/{fund_id}:
    get:
      operationId: fund_store.api.routes.get_fund
      tags:
        - Funds
      parameters:
        - $ref: "components.yml#/components/parameters/fund_id"
        - name: language
          in: query
          schema:
            type: string
          required: false
        - name: use_short_name
          in: query
          schema:
            type: boolean
          required: false
      responses:
        200:
          description: "If the fund exists then the data is returned."
          content:
            application/json:
              schema:
                $ref: "components.yml#/components/schemas/Fund"
        404:
          description: "Fund not found"
          content:
            texts/plain:
              schema:
                $ref: "components.yml#/components/schemas/Error"
  /funds/{fund_id}/rounds/{round_id}:
    get:
      tags:
        - Rounds
      summary: Returns the data on a specified round for a specific fund.
      description: Given a fund ID and a round ID we return the relavant round data.
      operationId: fund_store.api.routes.get_round
      parameters:
        - $ref: "components.yml#/components/parameters/fund_id"
        - $ref: "components.yml#/components/parameters/round_id"
        - name: language
          in: query
          schema:
            type: string
        - name: use_short_name
          in: query
          schema:
            type: boolean
          required: false
      responses:
        200:
          description: "If the round exists then the data is returned."
          content:
            application/json:
              schema:
                $ref: "components.yml#/components/schemas/Round"
        404:
          description: "Round not found from given fund id and round id."
          content:
            application/json:
              schema:
                $ref: "components.yml#/components/schemas/Error"
...